[MNT] renamed unit containing Loggable to loggable to avoid shadowing std lib.
